requirements.txt
asgiref==3.9.1
autopep8==2.3.2
black==25.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
diff-match-patch==20241021
Django==5.2.4
django-baton==5.0.2
django-import-export==4.3.9
flake8==7.3.0
idna==3.10
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
pycodestyle==2.14.0
pyflakes==3.4.0
python-decouple==3.8
requests==2.32.4
sqlparse==0.5.3
tablib==3.8.0
urllib3==2.5.0
